requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
openai==1.54.0
aiohttp==3.10.10
APScheduler==3.10.4
pytz==2024.2
